chore(GenericInput): add component
This component was made to meet the criteria of eslint when
facing a problem of exceeding the number of lines in the
AddMovie component. It's abble to render input elements from
props since its respective component is set as a getInput
object property with its type as key and calling its for method.